requirements.txt
Django
psycopg2
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support